requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
alabaster==0.7.12
apiclient==1.0.4
appdirs==1.4.4
argon2-cffi==20.1.0
asciitree==0.3.3
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autoflake==1.4
Babel==2.9.0
backcall==0.2.0
bandit==1.7.0
black==20.8b1
bleach==3.2.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
coverage==5.3.1
cramjam==2.5.0
cycler==0.10.0
darglint==1.5.8
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
dlib==19.21.1
doc8==0.8.1
docutils==0.16
entrypoints==0.3
eradicate==1.0
fasteners==0.16
fastparquet==0.8.0
filelock==3.0.12
flake8==3.8.4
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.3.1
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
flatbuffers==1.12
fsspec==0.8.5
future==0.18.2
gast==0.3.3
GDAL==3.0.4
geohash2==1.1
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hyperopt==0.2.5
idna==2.10
idna-ssl==1.1.0
igdb-api-v4==0.3.3
imagesize==1.2.0
importlib-metadata==2.1.1
importlib-resources==4.1.1
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.10
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mercantile==1.1.6
mistune==0.8.4
monotonic==1.5
mtcnn==0.1.0
multidict==5.1.0
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.6
numcodecs==0.7.2
numpy==1.19.4
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.8.1
pbr==5.5.1
pep8-naming==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycrypt==0.7.2
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.3
pymongo==3.11.2
pyparsing==2.4.7
pyrameter==0.2.0.post2
pyrsistent==0.17.3
pytest==6.2.1
pytest-cov==2.10.1
python-dateutil==2.8.1
python-dev-tools==2020.9.10
pytorch-lightning==1.1.3
pytube==10.5.1
pytz==2020.5
pyupgrade==2.7.4
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
restructuredtext-lint==1.3.2
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
Send2Trash==1.5.0
shadho==0.3.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stevedore==3.3.0
tabulate==0.8.9
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testfixtures==6.17.0
testpath==0.4.4
threadpoolctl==2.1.0
tokenize-rt==4.0.0
toml==0.10.2
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tox==3.20.1
tox-travis==0.12
tqdm==4.55.1
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
wcwidth==0.2.5
webencodings==0.5.1
wemake-python-styleguide==0.14.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zarr==2.6.1
zipp==3.4.0
